Make some changes
hello ignacio, why would the user see this statement "Please click on the link that was sent to your e-mail address<br>to ensure you get verified"? How are they going to get the link for verification? is it by mail? could you explain you "endif" statement in your code, so that i can get a better understanding what string to put in. If you think it is okay you can accept the pull request.